# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Competency Platform REST API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: take, in: query, schema: { type: integer } }
        - { name: skip, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    put:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/profile:
    get:
      operationId: jobProfile
      tags: [jobs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProfile"
...
components:
  schemas:
...
    Competency:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        level: { type: string }
        description: { type: ["string", "null"] }
        indicators: { type: ["string", "null"] }
        isShared: { type: boolean }
        sharedKey: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, type, level, isShared]
...
    Job:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        contractType: { type: ["string", "null"] }
        experienceLevel: { type: ["string", "null"] }
        departmentId: { type: ["string", "null"] }
        gradeId: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        departmentName: { type: ["string", "null"] }
        gradeName: { type: ["string", "null"] }
        competencyIds:
          type: array
          items: { type: string }
        competencyCount: { type: integer }
      required: [id, name]
    JobInput:
      type: object
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
        contractType: { type: string }
        experienceLevel: { type: ["string", "null"] }
        departmentId: { type: ["string", "null"] }
        gradeId: { type: ["string", "null"] }
        reportsToJobId: { type: ["string", "null"] }
        competencyIds:
          type: array
          items: { type: string }
      required: [name]
    JobProfile:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/Job"
        competencies:
          type: array
          items:
            $ref: "#/components/schemas/Competency"
        kpiSet:
          oneOf:
            - $ref: "#/components/schemas/KpiSet"
            - type: "null"
      required: [job, competencies]
...
    Kpi:
      type: object
      properties:
        id: { type: string }
        kpiGroupId: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        measure: { type: ["string", "null"] }
        target: { type: ["string", "null"] }
        frequency: { type: ["string", "null"] }
        weight: { type: ["number", "null"] }
        order: { type: integer }
      required: [id, name]
    KpiGroup:
      type: object
      properties:
        id: { type: string }
        kpiSetId: { type: string }
        competencyName: { type: string }
        compType: { type: ["string", "null"] }
        order: { type: integer }
        kpis:
          type: array
          items:
            $ref: "#/components/schemas/Kpi"
      required: [id, competencyName, kpis]
    KpiSet:
      type: object
      properties:
        id: { type: string }
        jobId: { type: string }
        summary: { type: ["string", "null"] }
        isAiGenerated: { type: boolean }
        groups:
          type: array
          items:
            $ref: "#/components/schemas/KpiGroup"
      required: [id, jobId, groups]
...